+++ HISTORY.md
@@ -7,6 +7,7 @@ History
 * Added `StateT` extension `liftR` to `ReactST` and deprecated `runState` methods handling `StateT` directly.
 * Added `colspan` and `rowspan` attributes.
 * Added `ReactS.zoom` and `ReactS.zoomU`.
+* Params for `ReactS.callback{,T}` no longer curried. ← Breaks backwards compatibility. Sorry. Are there more than 3 instances of this in the wild anyway?
 
 ### 0.5.1 ([commit log](https://github.com/japgolly/scalajs-react/compare/v0.5.0...v0.5.1))